// src/pages/GraphAndGlobe/components/CategoryChart/interfaces.ts
export interface ICityCategoryData {
  City: string;
  Country: string;
  "Urban Resilience Index": string;
  Environmental: number;
  Social: number;
  Governmental: number;
  "Air pollution": number;
  "Water pollution": number;
  "Land use and waste": number;
  "Natural disaster hazards": number;
  "Water availability": number;
  "Food availability": number;
  "Energy availability": number;
  Health: number;
  Education: number;
  Wealth: number;
  Safety: number;
  Livability: number;
  Transport: number;
  Inclusion: number;
  Accountability: number;
  "City budget": number;
  SME: number;
  "Public services": number;
  "Sustainability commitment": number;
}